Add support for stm32h7b3xx MCU's flash
