+++ arch/arm/src/stm32h7/hardware/stm32_sdmmc.h
@@ -30,6 +30,8 @@
 
 #if defined(CONFIG_STM32H7_STM32H7X3XX)
 #  include "stm32h7x3xx_sdmmc.h"
+#elif defined(CONFIG_STM32H7_STM32H7B3XX)
+#  include "stm32h7x3xx_sdmmc.h"
 #elif defined(CONFIG_STM32H7_STM32H7X7XX)
 #  include "stm32h7x3xx_sdmmc.h"
 #else
