+++ arch/arm/src/stm32h7/hardware/stm32_syscfg.h
@@ -30,6 +30,8 @@
 
 #if defined(CONFIG_STM32H7_STM32H7X3XX)
 #  include "hardware/stm32h7x3xx_syscfg.h"
+#elif defined(CONFIG_STM32H7_STM32H7B3XX)
+#  include "hardware/stm32h7x3xx_syscfg.h"
 #elif defined(CONFIG_STM32H7_STM32H7X7XX)
 #  include "hardware/stm32h7x3xx_syscfg.h"
 #else
